Use asio instead of WinSocks

+++ shared/src/socket_messages/identify_message.cpp
@@ -1,5 +1,6 @@
 #include "socket_messages/identify_message.h"
 
+#include <algorithm>
 #include <cstdint>
 #include <string>
 #include <vector>
@@ -47,7 +48,7 @@ namespace pine::socket_messages
 		return buffer;
 	}
 
-	size_t identify_message::get_body_size() const
+	uint64_t identify_message::get_body_size() const
 	{
 		return sizeof(uint8_t) + username.size();
 	}
